Support for attributes dir; refactoring
