+++ WebApplication/Initializer.cs
@@ -44,6 +44,8 @@ namespace WebApplication
             await _forge.CreateBucket(_resourceProvider.BucketName);
             _logger.LogInformation($"Bucket {_resourceProvider.BucketName} created");
 
+            var arranger = new Arranger(_forge, _resourceProvider.BucketName);
+
             // download default project files from the public location
             // specified by the appsettings.json
             using (var client = new HttpClient())
@@ -54,8 +56,6 @@ namespace WebApplication
                 {
                     throw new Exception("Default projects are not in sync with TLA names");
                 }
-
-                var bucketKey = _resourceProvider.BucketName;
 
                 for (var i = 0; i < defaultProjects.Length; i++)
                 {
@@ -74,26 +74,11 @@ namespace WebApplication
                         _logger.LogInformation("Upload to the app bucket");
 
                         Stream stream = await response.Content.ReadAsStreamAsync();
-                        await _forge.UploadObject(bucketKey, stream, project.OSSSourceModel);
+                        await _forge.UploadObject(_resourceProvider.BucketName, stream, project.OSSSourceModel);
                     }
-
-                    var documentUrl = await _forge.CreateSignedUrl(bucketKey, project.OSSSourceModel);
 
                     _logger.LogInformation("Adopt the project");
-                    var arranger = new Arranger(_forge, bucketKey);
-
-                    var adoptionData = await arranger.ForAdoption(documentUrl, tlaFilename);
-
-                    var status = await _fdaClient.Adopt(adoptionData); // ER: think - it's a business logic, so it might not deal with low-level WI and status
-                    if (status.Status != Status.Success)
-                    {
-                        _logger.LogError($"Failed to adopt {projectUrl}");
-                    }
-                    else
-                    {
-                        // rearrange generated data according to the parameters hash
-                        await arranger.Do(project);
-                    }
+                    await Adopt(arranger, project, tlaFilename);
                 }
             }
 
@@ -115,6 +100,27 @@ namespace WebApplication
 
             // delete bundles and activities
             await _fdaClient.CleanUp();
+        }
+
+        /// <summary>
+        /// Adapt the project.
+        /// </summary>
+        private async Task Adopt(Arranger arranger, Project project, string tlaFilename)
+        {
+            var inputDocUrl = await _forge.CreateSignedUrl(_resourceProvider.BucketName, project.OSSSourceModel);
+
+            var adoptionData = await arranger.ForAdoption(inputDocUrl, tlaFilename);
+
+            var status = await _fdaClient.Adopt(adoptionData); // ER: think: it's a business logic, so it might not deal with low-level WI and status
+            if (status.Status != Status.Success)
+            {
+                _logger.LogError($"Failed to adopt {project.Name}");
+            }
+            else
+            {
+                // rearrange generated data according to the parameters hash
+                await arranger.Do(project);
+            }
         }
     }
 }
